renamed ProcessUnitTest to ProcessIntegrationTest Added light-weight process unit tests
